refactor(synthetics): Move structs into implementing files
