+++ pkg/synthetics/monitors.go
@@ -9,6 +9,75 @@ const (
 	listMonitorsLimit = 100
 )
 
+// Monitor represents a New Relic Synthetics monitor.
+type Monitor struct {
+	ID           string            `json:"id,omitempty"`
+	Name         string            `json:"name"`
+	Type         MonitorType       `json:"type"`
+	Frequency    uint              `json:"frequency"`
+	URI          string            `json:"uri"`
+	Locations    []string          `json:"locations"`
+	Status       MonitorStatusType `json:"status"`
+	SLAThreshold float64           `json:"slaThreshold"`
+	UserID       uint              `json:"userId,omitempty"`
+	APIVersion   string            `json:"apiVersion,omitempty"`
+	ModifiedAt   *Time             `json:"modifiedAt,omitempty"`
+	CreatedAt    *Time             `json:"createdAt,omitempty"`
+	Options      MonitorOptions    `json:"options,omitempty"`
+}
+
+// MonitorScriptLocation represents a New Relic Synthetics monitor script location.
+type MonitorScriptLocation struct {
+	Name string `json:"name"`
+	HMAC string `json:"hmac"`
+}
+
+// MonitorScript represents a New Relic Synthetics monitor script.
+type MonitorScript struct {
+	Text      string                  `json:"scriptText"`
+	Locations []MonitorScriptLocation `json:"scriptLocations"`
+}
+
+// MonitorType represents a Synthetics monitor type.
+type MonitorType string
+
+// MonitorStatusType represents a Synthetics monitor status type.
+type MonitorStatusType string
+
+// MonitorOptions represents the options for a New Relic Synthetics monitor.
+type MonitorOptions struct {
+	ValidationString       string `json:"validationString,omitempty"`
+	VerifySSL              bool   `json:"verifySSL,omitempty"`
+	BypassHEADRequest      bool   `json:"bypassHEADRequest,omitempty"`
+	TreatRedirectAsFailure bool   `json:"treatRedirectAsFailure,omitempty"`
+}
+
+var (
+	// MonitorTypes specifies the possible types for a Synthetics monitor.
+	MonitorTypes = struct {
+		Ping            MonitorType
+		Browser         MonitorType
+		ScriptedBrowser MonitorType
+		APITest         MonitorType
+	}{
+		Ping:            "SIMPLE",
+		Browser:         "BROWSER",
+		ScriptedBrowser: "SCRIPT_BROWSER",
+		APITest:         "SCRIPT_API",
+	}
+
+	// MonitorStatus specifies the possible Synthetics monitor status types.
+	MonitorStatus = struct {
+		Enabled  MonitorStatusType
+		Muted    MonitorStatusType
+		Disabled MonitorStatusType
+	}{
+		Enabled:  "ENABLED",
+		Muted:    "MUTED",
+		Disabled: "DISABLED",
+	}
+)
+
 // ListMonitors is used to retrieve New Relic Synthetics monitors.
 func (s *Synthetics) ListMonitors() ([]*Monitor, error) {
 	resp := listMonitorsResponse{}
